+++ src/directives/ellipsisTooltip/ellipsisTooltip.directive.ts
@@ -1,0 +1,134 @@
+import {Directive, ElementRef, Input, OnChanges, OnDestroy, Renderer2, SimpleChanges, booleanAttribute, inject} from '@angular/core';
+import {nameof} from '@jscrpt/common';
+
+import {TooltipDirective} from '../../modules/tooltip';
+
+/**
+ * Directive that is used for displaying text from attached element in tooltip
+ */
+@Directive(
+{
+    selector: '[ellipsisTooltip]',
+    standalone: true,
+    hostDirectives:
+    [
+        TooltipDirective,
+    ],
+})
+export class EllipsisTooltipSADirective implements OnChanges, OnDestroy
+{
+    //######################### protected fields #########################
+
+    /**
+     * Instance of renderer
+     */
+    protected renderer: Renderer2 = inject(Renderer2);
+
+    /**
+     * Instance of tooltip directive
+     */
+    protected tooltip: TooltipDirective = inject(TooltipDirective);
+
+    /**
+     * HTML element from which is text taken for tooltip
+     */
+    protected elementValue: HTMLElement = inject(ElementRef).nativeElement;
+
+    /**
+     * Instance of mutation observer used for watching 
+     */
+    protected textObserver: MutationObserver|undefined|null = new MutationObserver(() => this.tooltip.tooltip = this.allowHtml ? this.element.innerHTML : this.element.innerText);
+
+    /**
+     * Original css class
+     */
+    protected originalCssClass: string|undefined|null;
+
+    //######################### public properties - inputs #########################
+
+    /**
+     * Css class appliet to element
+     */
+    @Input()
+    public ellipsisClass: string = 'text-ellipsis';
+
+    /**
+     * Indication whether are html tags allowed in tooltip text
+     */
+    @Input({transform: booleanAttribute})
+    public allowHtml: boolean = false;
+
+    /**
+     * Gets or sets HTML element from which is text taken for tooltip
+     */
+    @Input('ellipsisTooltip')
+    public get element(): HTMLElement
+    {
+        return this.elementValue;
+    }
+    public set element(value: HTMLElement|ElementRef<HTMLElement>)
+    {
+        if(!value)
+        {
+            return;
+        }
+
+        if(value instanceof ElementRef)
+        {
+            this.elementValue = value.nativeElement;
+
+            return;
+        }
+
+        this.elementValue = value;
+    }
+
+    //######################### public methods - implementation of OnChanges #########################
+    
+    /**
+     * Called when input value changes
+     */
+    public ngOnChanges(changes: SimpleChanges): void
+    {
+        if(nameof<EllipsisTooltipSADirective>('allowHtml') in changes)
+        {
+            this.tooltip.allowHtml = this.allowHtml;
+        }
+
+        if(nameof<EllipsisTooltipSADirective>('element') in changes ||
+           nameof<EllipsisTooltipSADirective>('ellipsisClass') in changes)
+        {
+            if(this.originalCssClass != this.ellipsisClass)
+            {
+                if(this.originalCssClass)
+                {
+                    this.renderer.removeClass(this.element, this.originalCssClass);
+                }
+    
+                this.renderer.addClass(this.element, this.ellipsisClass);
+                this.originalCssClass = this.ellipsisClass;
+            }
+
+            this.textObserver?.disconnect();
+            this.textObserver?.observe(this.element, {characterData: true, subtree: true, childList: true});
+        }
+    }
+
+    //######################### public methods - implementation of OnDestroy #########################
+    
+    /**
+     * Called when component is destroyed
+     */
+    public ngOnDestroy(): void
+    {
+        this.textObserver?.disconnect();
+        this.textObserver = null;
+    }
+
+    //######################### ng language server #########################
+    
+    /**
+     * Custom input type for `element` input
+     */
+    public static ngAcceptInputType_element: HTMLElement|ElementRef|'';
+}